[♻️ refactor] 토큰 인증방식 변경

+++ src/main/java/org/noostak/infra/S3ServiceImpl.java
@@ -26,7 +26,7 @@ public class S3ServiceImpl implements S3Service {
     public KeyAndUrl uploadImage(S3DirectoryPath dirPath, MultipartFile image) {
         try {
             if (image ==null || image.isEmpty()){
-                return null;
+                return KeyAndUrl.of(null,null);
             }
 
             return s3Storage.upload(dirPath, image);
@@ -37,12 +37,19 @@ public class S3ServiceImpl implements S3Service {
 
     @Override
     public String getImageUrl(String key) {
+        if (key ==null || key.isEmpty()){
+            return null;
+        }
+
         return s3Storage.findPublicUrlByKey(key);
     }
 
     @Override
     public void deleteImage(String key) {
         try {
+            if(key == null){
+                return;
+            }
             s3Storage.delete(key);
         } catch (Exception e) {
             throw new S3DeleteException(S3DeleteErrorCode.IMAGE_DELETE_FAILED,e.getMessage());
